internal/dsl: extract the shared predicate and path evaluator
The verification subsystem needs the same operator set and path walker the
detection engine uses, but not its Block combinator, which is bound to
detect's own types. Move the leaf-predicate evaluator and GetPath into
internal/dsl and leave evaluateBlock behind.
